+++ TravelingSalesmanProblem.h
@@ -72,6 +72,7 @@ Tour* AntColony_FindTour(const Graph* g);
 Tour* GeneticAlgorithm_FindTour(const Graph* g);
 Tour* TabuSearch_FindTour(const Graph* g);
 Tour* GRASP_FindTour(const Graph* g);
+Tour* ISPO_FindTour(const Graph* g);
 
 // In the future, it could be interesting implementing ISPO:
 // https://www.researchgate.net/publication/271285365_ISPO_A_New_Way_to_Solve_Traveling_Salesman_Problem